+++ spec/initializers/array_spec.rb
@@ -23,4 +23,20 @@ describe Array do
       expect(%w[foo bar baz].or).to eq 'foo, bar or baz'
     end
   end
+  context '.same?' do
+    it 'empty' do
+      expect([].same?).to be false
+    end
+    it 'one' do
+      expect([1].same?).to be true
+    end
+    it 'two' do
+      expect([1, 1].same?).to be true
+      expect([1, 2].same?).to be false
+    end
+    it 'three' do
+      expect([2, 2, 2].same?).to be true
+      expect([1, 2, 3].same?).to be false
+    end
+  end
 end